set the reserve pool to 0
We are changing this as it allows us to lock in the max connections per pgbouncer instance by using `default_pool_size` and is one less configuration to have to set/manage.

Reserve Pool Size was defaulted to 1 but we are changing it to default to 0.

The way Reserve Pool Size works is that if a client is not serviced within the `reserve_pool_timeout` of 5 seconds (default), a connection in the reserve pool would be opened (if enabled).